+++ CHANGELOG.md
@@ -7,19 +7,19 @@ The format is based on [Keep a Changelog]() and this project adheres to
 
 ## [Unreleased]
 
-## [2017-03-02]
+## [2017-03-02] 0.0.4
 
 ### Added
 
 - Working (but unoptimized) Timsort for full sorting.
 
-## [2017-02-28]
+## [2017-02-28] 0.0.3
 
 ### Added
 
 - Nearly working Timsort for full sorting.
 
-## [2017-02-26]
+## [2017-02-26] 0.0.2
 
 ### Added